Remove Sentence type.
Sentence is now just an alias for Vec<Token>. The original motivation
for adding Sentence as a wrapper around Vec<Token> was to implement
the Display trait, such that Sentence instances could be written in
CoNLL-X format.

However, I ended up implementing/delegating more and more of the Vec
methods. Since automatic delegation is not supported (yet) and Deref
should really only be used for smart pointers, I decided to just use
Vec<Token> throughout this crate to represent sentences. There is a
small wrapper type, DisplaySentence, specifically for the purpose of
formatting CoNLL-X-style output.

This change breaks the API for all users of this crate. But we're not
post 1.0 yet :).