Rewrite large parts of conllx to provide a graph representation.
Many downstream uses of conllx consist of converting the CoNLL-X
tokens to a proper graph. This change makes a graph representation the
default representation.

The new `Sentence` data type stores tokens and edges in a petgraph
DiGraph.  A simple API is provided for using the (non-)projective
graphs. For advanced uses, the inner petgraph can be unwrapped and
used.