Represent features as a BTreeMap
This removes the dual String/BTreeMap representation of features,
where the map was lazily initialized.

The dual representation setup was clunky and made it impossible to
change the feature map in-place. This representation was initially
chosen to avoid the feature parsing overhead. However, the difference
in performance is fairly small:

Old: 4.446
New: 4.776